fav button & shortcut implemented in preview
